Migrate fjs/types/object to .f.mjs, split types into types.ts
Splits the type-level API (OptionalMap, RequiredMap, StringMap, Entry,
OneKey, NotUnion, SingleProperty) into a sibling types.ts, and converts
the runtime helpers (at, sort, fromEntries, fromMap, isObject,
definedValues, definedEntries) to JSDoc-typed module.f.mjs, following
the convention introduced by #1481/#1483. Updates all dependents to
import runtime values from module.f.mjs and types from types.ts,
splitting mixed import statements where needed.
