+++ fjs/bnf/descent/module.f.ts
@@ -16,7 +16,7 @@
 import type { CodePoint } from '../../text/utf16/types.ts'
 import { rangeDecode, type TerminalRange } from '../module.f.ts'
 import { contains as rangeContains } from '../../types/range/module.f.mjs'
-import { definedEntries } from '../../types/object/module.f.ts'
+import { definedEntries } from '../../types/object/module.f.mjs'
 import { emptyTagMap, toData, type Rule as DataRule, type Sequence } from '../data/module.f.ts'
 import { type Rule as FRule } from '../module.f.ts'
