+++ fjs/bnf/ll1/module.f.ts
@@ -17,7 +17,8 @@ import type { RangeMapArray } from '../../types/range_map/types.ts'
 import { rangeMap } from '../../types/range_map/module.f.mjs'
 import { contains, set, type StringSet } from '../../types/string_set/module.f.ts'
 import { rangeDecode } from '../module.f.ts'
-import { definedEntries, type StringMap } from '../../types/object/module.f.ts'
+import { definedEntries } from '../../types/object/module.f.mjs'
+import type { StringMap } from '../../types/object/types.ts'
 import { emptyTagMap, type EmptyTag, type RuleSet, toData } from '../data/module.f.ts'
 import { type Rule as FRule } from '../module.f.ts'
