+++ fjs/cas/evo/module.f.ts
@@ -53,7 +53,8 @@ import { decodeText, encodeText, dialect, checkReferences, isHash, type LockMap,
 import type { Ok, Result } from '../../types/result/types.ts'
 import { ok, error } from '../../types/result/module.f.mjs'
 import { nonEmpty, empty as elEmpty } from '../../effects/list/module.f.ts'
-import { at, definedEntries, type StringMap } from '../../types/object/module.f.ts'
+import { at, definedEntries } from '../../types/object/module.f.mjs'
+import type { StringMap } from '../../types/object/types.ts'
 import { unwrap } from '../../types/nullable/module.f.mjs'
 import type { Vec } from '../../types/bit_vec/types.ts'
 import { isNotFound, type IoResult } from '../../effects/node/module.f.ts'
