+++ fjs/dev/module.f.ts
@@ -14,7 +14,7 @@ import {
     type Readdir
 } from '../effects/node/module.f.ts'
 import { cmp as strCmp } from '../types/string/module.f.mjs'
-import type { StringMap } from '../types/object/module.f.ts'
+import type { StringMap } from '../types/object/types.ts'
 import { unwrap } from '../types/result/module.f.mjs'
 import { pure, step, type Effect } from '../effects/module.f.ts'
 import { join, relativize, toPosix } from '../path/module.f.ts'
