+++ fjs/djs/ast/module.f.ts
@@ -7,7 +7,7 @@ import type { Primitive, Array, Unknown } from '../module.f.ts'
 import type { List } from '../../types/list/types.ts'
 import { concat, fold, last, map, take, toArray } from '../../types/list/module.f.mjs'
 import type { Entry } from '../../types/ordered_map/types.ts'
-import { fromEntries } from '../../types/object/module.f.ts'
+import { fromEntries } from '../../types/object/module.f.mjs'
 
 const { entries } = Object
