+++ fjs/djs/ast/proof.f.ts
@@ -1,4 +1,4 @@
-import { sort } from '../../types/object/module.f.ts'
+import { sort } from '../../types/object/module.f.mjs'
 import { run } from './module.f.ts'
 import { stringifyAsTree } from '../serializer/module.f.ts'
 import { assertEq } from '../../asserts/module.f.mjs'
