+++ fjs/djs/parser/module.f.ts
@@ -11,7 +11,7 @@ import type { Fold } from '../../types/function/operator/types.ts'
 import type { DjsToken, DjsTokenWithMetadata } from '../tokenizer/module.f.ts'
 import { setReplace, at } from '../../types/ordered_map/module.f.mjs'
 import type { OrderedMap } from '../../types/ordered_map/types.ts'
-import { fromMap } from '../../types/object/module.f.ts'
+import { fromMap } from '../../types/object/module.f.mjs'
 import type { AstArray, AstConst, AstModule, AstModuleRef } from '../ast/module.f.ts'
 import type { TokenMetadata } from '../../js/tokenizer/module.f.ts'
 import { assertEq } from '../../asserts/module.f.mjs'
