+++ fjs/djs/parser/proof.f.ts
@@ -1,7 +1,7 @@
 import { parseFromTokens } from './module.f.ts'
 import { tokenize, type DjsTokenWithMetadata } from '../tokenizer/module.f.ts'
 import { toArray } from '../../types/list/module.f.mjs'
-import { sort } from '../../types/object/module.f.ts'
+import { sort } from '../../types/object/module.f.mjs'
 import { stringToList } from '../../text/utf16/module.f.mjs'
 import { stringifyAsTree } from '../serializer/module.f.ts'
 import { stringify } from '../../media/json/module.f.ts'
