+++ fjs/djs/tokenizer/module.f.ts
@@ -59,7 +59,7 @@ import { contains } from "../../types/range/module.f.mjs"
 import type { List } from "../../types/list/types.ts"
 import { concat, empty, filter, flat, flatMap, fold, map, stateScan, toArray } from "../../types/list/module.f.mjs"
 import { stringifyAsTree } from "../serializer/module.f.ts"
-import { sort } from "../../types/object/module.f.ts"
+import { sort } from "../../types/object/module.f.mjs"
 import type { Unknown } from "../module.f.ts"
 
 // Builds the single-token grammar that jsGrammar's whole-file `tokens` rule repeats.
