+++ fjs/effects/node/module.ts
@@ -43,7 +43,7 @@ import type { Result } from '../../types/result/types.ts'
 import { error, ok } from '../../types/result/module.f.mjs'
 import { asyncTryCatch } from '../../types/result/module.ts'
 import { fromVec, listToVec, toVec } from '../../types/uint8array/module.f.mjs'
-import type { StringMap } from '../../types/object/module.f.ts'
+import type { StringMap } from '../../types/object/types.ts'
 import { maxLengthBytes } from '../../types/bit_vec/module.f.mjs'
 
 type Server = {
