+++ fjs/fsm/module.f.ts
@@ -5,7 +5,7 @@
  */
 import type { List } from '../types/list/types.ts'
 import { equal, isEmpty, fold, toArray, scan, foldScan, empty as emptyList } from '../types/list/module.f.mjs'
-import type { StringMap } from '../types/object/module.f.ts'
+import type { StringMap } from '../types/object/types.ts'
 import { toRangeMap, union as byteSetUnion, one, empty, range, type ByteSet } from '../types/byte_set/module.f.ts'
 import { intersect, type SortedSet, union as sortedSetUnion } from '../types/sorted_set/module.f.ts'
 import type { RangeMap, Properties, RangeMapArray, Entry } from '../types/range_map/types.ts'
