+++ fjs/js/tokenizer/proof.f.ts
@@ -1,7 +1,7 @@
 import { tokenize, type JsToken, type JsTokenWithMetadata } from './module.f.ts'
 import { map, toArray } from '../../types/list/module.f.mjs'
 import { stringifyAsTree } from '../../djs/serializer/module.f.ts'
-import { sort } from '../../types/object/module.f.ts'
+import { sort } from '../../types/object/module.f.mjs'
 import { stringToList } from '../../text/utf16/module.f.mjs'
 
 const tokenizeString
