+++ fjs/media/html/module.f.ts
@@ -7,7 +7,8 @@
 import type { List } from '../../types/list/types.ts'
 import { map, flatMap, flat, concat as listConcat } from '../../types/list/module.f.mjs'
 import { concat, concat as stringConcat } from '../../types/string/module.f.mjs'
-import { definedEntries, type Entry, type StringMap } from '../../types/object/module.f.ts'
+import { definedEntries } from '../../types/object/module.f.mjs'
+import type { Entry, StringMap } from '../../types/object/types.ts'
 import { compose } from '../../types/function/module.f.mjs'
 import { stringToList } from '../../text/utf16/module.f.mjs'
 import { includes } from '../../types/array/module.f.mjs'
