+++ fjs/media/json/module.f.ts
@@ -18,7 +18,8 @@ import { stringToList } from '../../text/utf16/module.f.mjs'
 import type { Result } from '../../types/result/types.ts'
 import { parse as parseTokens } from './parser/module.f.ts'
 import { tokenize } from './tokenizer/module.f.ts'
-import { at, definedEntries, type Entry as ObjectEntry } from '../../types/object/module.f.ts'
+import { at, definedEntries } from '../../types/object/module.f.mjs'
+import type { Entry as ObjectEntry } from '../../types/object/types.ts'
 import { compose, fn } from '../../types/function/module.f.mjs'
 import { objectWrap, arrayWrap, stringSerialize, numberSerialize, nullSerialize, boolSerialize } from './serializer/module.f.ts'
 import { boolean as rttiBoolean, number as rttiNumber, string as rttiString, or, record, array as rttiArray } from '../../types/rtti/module.f.ts'
