+++ fjs/media/json/tokenizer/proof.f.ts
@@ -1,7 +1,7 @@
 import { tokenize, type JsonToken } from './module.f.ts'
 import { toArray } from '../../../types/list/module.f.mjs'
 import { stringifyAsTree } from '../../../djs/serializer/module.f.ts'
-import { sort } from '../../../types/object/module.f.ts'
+import { sort } from '../../../types/object/module.f.mjs'
 import { stringToList } from '../../../text/utf16/module.f.mjs'
 
 const tokenizeString
