+++ fjs/text/utf8/proof.f.ts
@@ -1,6 +1,6 @@
 import { toCodePointList, fromCodePointList, fromVec, utf8ByteToCodePointOp } from './module.f.mjs'
 import { stringify as jsonStringify } from '../../media/json/module.f.ts'
-import { sort } from '../../types/object/module.f.ts'
+import { sort } from '../../types/object/module.f.mjs'
 import { toArray } from '../../types/list/module.f.mjs'
 import { msb, u8ListToVec, vec } from '../../types/bit_vec/module.f.mjs'
 import { assertEq } from '../../asserts/module.f.mjs'
