+++ fjs/types/array/proof.f.ts
@@ -1,6 +1,6 @@
 import { at, first, last, head, tail, splitFirst, splitLast, empty } from './module.f.mjs'
 import { stringify as jsonStringify } from '../../media/json/module.f.ts'
-import { sort } from '../object/module.f.ts'
+import { sort } from '../object/module.f.mjs'
 import { assertEq, assertNotNullish } from '../../asserts/module.f.mjs'
 
 const stringify = jsonStringify(sort)
