+++ fjs/types/btree/find/proof.f.ts
@@ -2,7 +2,7 @@ import type { Result } from './types.ts'
 import { find as btreeFind } from './module.f.mjs'
 import { map, toArray } from '../../list/module.f.mjs'
 import { stringify, type Unknown } from '../../../media/json/module.f.ts'
-import { sort } from '../../object/module.f.ts'
+import { sort } from '../../object/module.f.mjs'
 import type { TNode } from '../types/types.ts'
 import { cmp } from '../../string/module.f.mjs'
 import { set as setSet } from '../set/module.f.mjs'
