+++ fjs/types/range_map/proof.f.ts
@@ -1,7 +1,7 @@
 import type { RangeMapArray, Properties, RangeMap } from './types.ts'
 import { get, merge, fromRange, rangeMap } from './module.f.mjs'
 import { stringify, type Unknown } from '../../media/json/module.f.ts'
-import { sort } from '../object/module.f.ts'
+import { sort } from '../object/module.f.mjs'
 import { union, type SortedSet } from '../sorted_set/module.f.ts'
 import { equal, toArray } from '../list/module.f.mjs'
 import { strictEqual } from '../function/operator/module.f.mjs'
