+++ fjs/types/rtti/common/module.f.ts
@@ -41,7 +41,8 @@ import type { Error, Result as CommonResult } from '../../result/types.ts'
 import { error, ok } from '../../result/module.f.mjs'
 import type { Ts } from '../ts/module.f.ts'
 import { isArray as commonIsArray } from '../../array/module.f.mjs'
-import { isObject as commonIsObject, type StringMap } from '../../object/module.f.ts'
+import { isObject as commonIsObject } from '../../object/module.f.mjs'
+import type { StringMap } from '../../object/types.ts'
 
 /** A path to a sub-value within the validated structure. Each step is an object key or stringified array index. */
 export type Path = readonly string[]
