+++ fjs/types/rtti/ts/module.f.ts
@@ -12,7 +12,7 @@ import { primitive, union, printer as tsPrinter } from '../../ts/module.f.mjs'
 import type { Tag0, Tag1, Const, Or, String as RttiString, Struct, Tuple, Type, ConstObject } from '../module.f.ts'
 import type { Assert } from '../../../asserts/types.ts'
 import type { phantomKey } from '../../phantom/types.ts'
-import type { StringMap } from '../../object/module.f.ts'
+import type { StringMap } from '../../object/types.ts'
 
 /**
  * The set of primitive literal types representable as rtti `Const` values.
